requirements.txt
matplotlib==3.1.2
numpy==1.17.0
opencv_python==4.1.2.30
Pillow==10.1.0
scikit_image==0.17.2
scikit_learn==1.3.2
skimage==0.0
torch==1.10.0
torchmetrics==1.2.1
torchvision==0.11.1
tqdm==4.60.0
